+++ mir-module/src/main/java/org/mycore/mir/authorization/accesskey/exception/MIRAccessKeyCollisionException.java
@@ -21,10 +21,10 @@
  * 59 Temple Place - Suite 330, Boston, MA  02111-1307 USA
  */
 
-package org.mycore.mir.authorization.accesskeys.exceptions;
+package org.mycore.mir.authorization.accesskey.exception;
 
-public class MIRAccessKeyNotFoundException extends MIRAccessKeyException {
-    public MIRAccessKeyNotFoundException(String errorMessage) {
+public class MIRAccessKeyCollisionException extends MIRAccessKeyException {
+    public MIRAccessKeyCollisionException(String errorMessage) {
         super(errorMessage);
     }
 }
