+++ mir-module/src/main/java/org/mycore/mir/authorization/accesskey/exception/MIRAccessKeyException.java
@@ -21,10 +21,13 @@
  * 59 Temple Place - Suite 330, Boston, MA  02111-1307 USA
  */
 
-package org.mycore.mir.authorization.accesskeys.exceptions;
+package org.mycore.mir.authorization.accesskey.exception;
 
-public class MIRAccessKeyCollisionException extends MIRAccessKeyException {
-    public MIRAccessKeyCollisionException(String errorMessage) {
+import org.mycore.common.MCRException;
+
+public class MIRAccessKeyException extends MCRException {
+    public MIRAccessKeyException(String errorMessage) {
         super(errorMessage);
     }
 }
+
