+++ mir-module/src/main/java/org/mycore/mir/authorization/accesskey/package-info.java
@@ -20,8 +20,5 @@
  * If not, write to the Free Software Foundation Inc.,
  * 59 Temple Place - Suite 330, Boston, MA  02111-1307 USA
  */
-/**
- * @author Ren\u00E9 Adler (eagle)
- *
- */
-package org.mycore.mir.authorization.accesskeys;
+
+package org.mycore.mir.authorization.accesskey;
