+++ mir-module/src/main/resources/META-INF/mir-module-mappings.xml
@@ -1,3 +1,3 @@
 <entity-mappings>
-  <entity class="org.mycore.mir.authorization.accesskeys.backend.MIRAccessKey" />
+  <entity class="org.mycore.mir.authorization.accesskey.backend.MIRAccessKey" />
 </entity-mappings>
