+++ mir-module/src/main/resources/config/mir/mycore.properties
@@ -15,7 +15,7 @@ MCR.LanguageClassification=rfc5646
 #MCR.CLI.Classes.External=%MCR.CLI.Classes.External%,org.mycore.mir.migration.MIRMigration04Utils
 #MCR.CLI.Classes.External=%MCR.CLI.Classes.External%,org.mycore.mir.migration.MIRMigration201603Utils
 #MCR.CLI.Classes.External=%MCR.CLI.Classes.External%,org.mycore.mir.migration.MIRMigration201706Utils
-MCR.CLI.Classes.External=%MCR.CLI.Classes.External%,org.mycore.mir.migration.MIRMigration202006Utils,org.mycore.mir.authorization.accesskeys.MIRAccessKeyUtils
+MCR.CLI.Classes.External=%MCR.CLI.Classes.External%,org.mycore.mir.migration.MIRMigration202006Utils,org.mycore.mir.authorization.accesskey.cli.MIRAccessKeyCommands
 
 MCR.ContentTransformer.mycoreobject-compress.Stylesheet=xsl/mods2mods.xsl,xsl/mods2dc.xsl
 
@@ -48,7 +48,7 @@ MCR.ContentTransformer.mets-dfg.Stylesheet=xsl/mets/mets-dfg.xsl
 ##############################################################################
 # Configure Hibernate Mappings                                               #
 ##############################################################################
-MCR.Hibernate.Mappings=%MCR.Hibernate.Mappings%,org.mycore.mir.authorization.accesskeys.backend.MIRAccessKey
+MCR.Hibernate.Mappings=%MCR.Hibernate.Mappings%,org.mycore.mir.authorization.accesskey.backend.MIRAccessKey
 
 
 #############################################################################
@@ -57,7 +57,7 @@ MCR.Hibernate.Mappings=%MCR.Hibernate.Mappings%,org.mycore.mir.authorization.acc
 MIR.AccessKeyStrategy.FallbackClass=org.mycore.mir.authorization.MIRStrategy
 # MIR.AccessKeyStrategy.ObjectTypes=mods,derivate
 
-MCR.URIResolver.ModuleResolver.accesskeys=org.mycore.mir.authorization.accesskeys.MIRAccessKeyResolver
+MCR.URIResolver.ModuleResolver.accesskeys=org.mycore.mir.authorization.accesskey.MIRAccessKeyResolver
 
 #############################################################################
 # HTMLEditor                                                                #
@@ -135,8 +135,8 @@ MCR.EventHandler.MCRDerivate.017.Class=org.mycore.datamodel.common.MCRServiceFla
 MCR.EventHandler.MCRObject.900.Class=org.mycore.wfc.mail.MCRMailEventHandler
 MCR.EventHandler.MCRDerivate.900.Class=org.mycore.wfc.mail.MCRMailEventHandler
 
-MCR.EventHandler.MCRObject.011.Class=org.mycore.mir.authorization.accesskeys.MIRAccessKeyEventHandler
-MCR.EventHandler.MCRDerivate.011.Class=org.mycore.mir.authorization.accesskeys.MIRAccessKeyEventHandler
+MCR.EventHandler.MCRObject.011.Class=org.mycore.mir.authorization.accesskey.MIRAccessKeyEventHandler
+MCR.EventHandler.MCRDerivate.011.Class=org.mycore.mir.authorization.accesskey.MIRAccessKeyEventHandler
 
 
 ##############################################################################
@@ -436,7 +436,7 @@ MCR.ContentTransformer.mods2csv.FileExtension=csv
 ##############################################################################
 # Jersey Resources                                                           #
 ##############################################################################
-MCR.Jersey.Resource.Packages=%MCR.Jersey.Resource.Packages%,org.mycore.mir.statistics,org.mycore.mir.statistics,org.mycore.mir.authorization.accesskeys
+MCR.Jersey.Resource.Packages=%MCR.Jersey.Resource.Packages%,org.mycore.mir.statistics,org.mycore.mir.statistics,org.mycore.mir.authorization.frontend.accesskey
 
 
 ##############################################################################
